Separate hashi shim interfaces so they can be loaded independently of implementations. Use code splitting so xAPI and H5P code bloat doesn't affect every hashi load.
