+++ packages/hashi/src/H5P/loadBinary.js
